Fixed llvm intrinsics memset and memmove.

+++ etc/libbambu/memmove.c
@@ -44,7 +44,7 @@ _llvm_memmove_p0i8_p0i8_i32(void *dest, const void *src, size_t len, unsigned in
 }
 
 void
-_llvm_memmove_p0i8_p0i8_i64(void *dest, const void *src, size_t len, unsigned long long int align, _Bool isvolatile)
+_llvm_memmove_p0i8_p0i8_i64(void *dest, const void *src, unsigned long long int len, unsigned int align, _Bool isvolatile)
 {
   char *d = dest;
   const char *s = src;
